+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 Notable changes to this project goes here
 
-## [v3.5.0] - unreleased
+## [v3.5.0] - 01.11.2025 
 
 ### Added